moved pages folder to app
